build fix

+++ src/test/java/com/example/io_backend/service/EquipmentServiceTest.java
@@ -36,7 +36,7 @@ class EquipmentServiceTest {
     @Test
     void getEquipmentById() {
 
-        when(equipmentRepository.findById(any(Long.class))).thenReturn(Optional.of(new Equipment(1L, "eq1", null)));
+        when(equipmentRepository.findById(any(Integer.class))).thenReturn(Optional.of(new Equipment(1, "eq1", null)));
 
         var result = equipmentService.getEquipmentById(1);
 
@@ -48,7 +48,7 @@ class EquipmentServiceTest {
     void getEquipmentByName() {
         var expected = new EquipmentResponse("eq1");
 
-        when(equipmentRepository.getEquipmentByName(any(String.class))).thenReturn(List.of(new Equipment(1L, "eq1", null)));
+        when(equipmentRepository.getEquipmentByName(any(String.class))).thenReturn(List.of(new Equipment(1, "eq1", null)));
 
         var result = equipmentService.getEquipmentByName("eq1");
 
@@ -76,7 +76,7 @@ class EquipmentServiceTest {
         var expected = new EquipmentResponse(dto.getName());
 
 
-        when(equipmentRepository.findById(any(Long.class))).thenReturn(Optional.of(new Equipment(id, "eq1", null)));
+        when(equipmentRepository.findById(any(Integer.class))).thenReturn(Optional.of(new Equipment(id, "eq1", null)));
         when(equipmentRepository.save(any(Equipment.class))).thenAnswer(x -> x.getArguments()[0]);
 
         var result = equipmentService.updateEquipment(dto, id);
@@ -97,9 +97,9 @@ class EquipmentServiceTest {
 
     private List<Equipment> getEquipmentTestData() {
         return  List.of(
-                new Equipment(1L, "eq1", null),
-                new Equipment(2L, "eq2", null),
-                new Equipment(3L, "eq3", null)
+                new Equipment(1, "eq1", null),
+                new Equipment(2, "eq2", null),
+                new Equipment(3, "eq3", null)
         );
     }
